feat: move compliance check to individual asset to reduce loading time
